requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
certifi==2018.4.16
more-itertools==4.2.0
pluggy==0.6.0
py==1.5.3
pytest==3.6.2
six==1.11.0
Django==2.2.4
djangorestframework==3.10.2
pytz==2019.3
sqlparse==0.3.0
pytest-django==3.6.0
pytest-cov==2.8.1
django-rest-auth
drf-extensions==0.5.0
requests==2.22.0
